panel: Start work on a C based panel implementation
Right now this is literally nothing more than a main entry point, and
removes the old panel build entirely. However it does let us start
working on a C based critical chain implementation, that isn't plagued
by the problems of the current Vala panel.

Note that this is *not* a "Budgie rewrite", this is the start of an
attempt to remove the old derpiness. Also, we import the known-to-work
Brisk helpers util.h to ensure we keep the code clean.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>